Operators: Cleanup registration process
